[MA-8] Separated functionality of Display between MainViewer and Surfer
