+++ applications/mne_analyze/extensions/mainviewer/centralview.cpp
@@ -47,12 +47,6 @@
 
 //*************************************************************************************************************
 //=============================================================================================================
-// INCLUDES
-//=============================================================================================================
-
-
-//*************************************************************************************************************
-//=============================================================================================================
 // QT INCLUDES
 //=============================================================================================================
 
@@ -94,7 +88,7 @@ CentralView::CentralView()
     : QWidget(),
       m_view3d_container(nullptr),
       m_view3d_gridlayout(nullptr),
-      m_pSurfaceMesh(new CustomMesh)
+      m_pRootEntity(new QEntity)
 {
     init();
 }
@@ -109,14 +103,8 @@ void CentralView::init()
     //QGridLayout is used so the container constantly resizes to the size of the QMdiSubwindow
     m_view3d_gridlayout = new QGridLayout(this);
 
-
     Qt3DExtras::Qt3DWindow *pWindow = new Qt3DExtras::Qt3DWindow();
-    QEntity *pRootEntity = createEntityTree();
-    pWindow->setRootEntity(pRootEntity);
-
-    //create lights
-    QEntity *pLightRoot = createLightEntity();
-    pLightRoot->setParent(pRootEntity);
+    pWindow->setRootEntity(m_pRootEntity);
 
     QPickingSettings *pPickSettings = pWindow->renderSettings()->pickingSettings();
     pPickSettings->setFaceOrientationPickingMode(QPickingSettings::FrontAndBackFace);
@@ -126,7 +114,7 @@ void CentralView::init()
     QCamera *pCamera = pWindow->camera();
     pCamera->setPosition(QVector3D(0,0,1));
 
-    Qt3DExtras::QFirstPersonCameraController *pCamController = new Qt3DExtras::QFirstPersonCameraController(pRootEntity);
+    Qt3DExtras::QFirstPersonCameraController *pCamController = new Qt3DExtras::QFirstPersonCameraController(m_pRootEntity);
     pCamController->setCamera(pCamera);
 
     //A container is created to contain the QWindow that comes from BrainView, then a minimum size is set
@@ -136,83 +124,4 @@ void CentralView::init()
     m_view3d_container->setFocusPolicy(Qt::TabFocus);
     //The loaded surfaces, as a QWindow is added to the created container
     m_view3d_gridlayout->addWidget(m_view3d_container);
-}
-
-
-//*************************************************************************************************************
-
-QEntity* CentralView::createEntityTree()
-{
-    QEntity *pRootEntity = new QEntity();
-
-    //Create mesh entity
-    QEntity *pMeshEntity = new QEntity(pRootEntity);
-    pMeshEntity->addComponent(m_pSurfaceMesh);
-
-    Qt3DExtras::QPerVertexColorMaterial *pMaterial = new Qt3DExtras::QPerVertexColorMaterial();
-    pMeshEntity->addComponent(pMaterial);
-
-    Qt3DCore::QTransform *pTransform = new Qt3DCore::QTransform();
-    pMeshEntity->addComponent(pTransform);
-
-    //Add object picker
-    QObjectPicker *pObjectPicker = new QObjectPicker;
-    pObjectPicker->setDragEnabled(false);
-    pObjectPicker->setHoverEnabled(false);
-    connect(pObjectPicker, &QObjectPicker::pressed,
-            this, &CentralView::userClicked);
-    pMeshEntity->addComponent(pObjectPicker);
-
-    //Create click point
-    QEntity *pSphereEntity = new QEntity(pRootEntity);
-    m_pointMesh = new Qt3DExtras::QSphereMesh;
-    m_pointMesh->setRadius(0.001f);
-    pSphereTransform = new Qt3DCore::QTransform();
-    pSphereEntity->addComponent(m_pointMesh);
-    pSphereEntity->addComponent(pSphereTransform);
-    Qt3DExtras::QPhongMaterial *pSphereMat = new Qt3DExtras::QPhongMaterial;
-    pSphereMat->setAmbient(Qt::red);
-    pSphereEntity->addComponent(pSphereMat);
-
-    return pRootEntity;
-}
-
-
-//*************************************************************************************************************
-
-QEntity* CentralView::createLightEntity()
-{
-    QEntity *pLightRoot = new QEntity();
-    //Setup light positions, intensities and color
-    QList<QVector3D> lLightPositions;
-    const QColor lightColor(255,255,255);
-    const float lightIntensity = 0.2f;
-
-    lLightPositions << QVector3D(-0.5,0,0) << QVector3D(0,0,-0.5)
-                    << QVector3D(0.5,0,0) << QVector3D(0,0,0.5)
-                    << QVector3D(0,0.5,0) << QVector3D(0,-0.5,0);
-
-    //Create all the lights - make it shine
-    for(int i = 0; i < lLightPositions.size(); ++i) {
-        //Light source
-        Qt3DCore::QEntity* pLightEntity = new Qt3DCore::QEntity(pLightRoot);
-
-        Qt3DCore::QTransform* pTransform = new Qt3DCore::QTransform();
-        pTransform->setTranslation(lLightPositions.at(i));
-        pLightEntity->addComponent(pTransform);
-
-        Qt3DRender::QPointLight *pPointLight = new Qt3DRender::QPointLight(pLightEntity);
-        pPointLight->setColor(lightColor);
-        pPointLight->setIntensity(lightIntensity);
-        pLightEntity->addComponent(pPointLight);
-    }
-    return pLightRoot;
-}
-
-
-//*************************************************************************************************************
-
-void CentralView::userClicked(QPickEvent *event)
-{
-
 }
