+++ applications/mne_analyze/extensions/surfer/surfer.cpp
@@ -44,6 +44,19 @@
 #include <anShared/Model/surfacemodel.h>
 #include <anShared/Data/surfacedata.h>
 #include <anShared/Management/analyzedata.h>
+#include "anShared/Utils/metatypes.h"
+#include <disp3D/engine/model/3dhelpers/custommesh.h>
+
+
+//*************************************************************************************************************
+//=============================================================================================================
+// QT INCLUDES
+//=============================================================================================================
+
+#include <Qt3DCore>
+#include <Qt3DRender>
+#include <Qt3DExtras>
+#include <QEntity>
 
 
 //*************************************************************************************************************
@@ -52,7 +65,10 @@
 //=============================================================================================================
 
 using namespace SURFEREXTENSION;
+using namespace DISP3DLIB;
 using namespace ANSHAREDLIB;
+using namespace Qt3DRender;
+using namespace Qt3DCore;
 
 
 //*************************************************************************************************************
@@ -62,7 +78,8 @@ using namespace ANSHAREDLIB;
 
 Surfer::Surfer()
 : m_pControl(Q_NULLPTR)
-, m_pView(Q_NULLPTR)
+, m_pSurfaceMesh(new CustomMesh)
+, m_pSurferRoot(new QEntity)
 {
 
 }
@@ -90,6 +107,61 @@ QSharedPointer<IExtension> Surfer::clone() const
 void Surfer::init()
 {
     m_pSurfaceModel = m_analyzeData->loadSurface(QStringLiteral("./MNE-sample-data/subjects/sample/surf/rh.pial"));
+    // initialize 3D stuff:
+    //Create mesh entity
+    QEntity *pMeshEntity = new QEntity(m_pSurferRoot);
+    pMeshEntity->addComponent(m_pSurfaceMesh);
+
+    Qt3DExtras::QPerVertexColorMaterial *pMaterial = new Qt3DExtras::QPerVertexColorMaterial();
+    pMeshEntity->addComponent(pMaterial);
+
+    Qt3DCore::QTransform *pTransform = new Qt3DCore::QTransform();
+    pMeshEntity->addComponent(pTransform);
+
+    //Add object picker
+    QObjectPicker *pObjectPicker = new QObjectPicker;
+    pObjectPicker->setDragEnabled(false);
+    pObjectPicker->setHoverEnabled(false);
+    connect(pObjectPicker, &QObjectPicker::pressed,
+            this, &Surfer::onClick);
+    pMeshEntity->addComponent(pObjectPicker);
+
+    //Create click point
+    QEntity *pSphereEntity = new QEntity(m_pSurferRoot);
+    m_pointMesh = new Qt3DExtras::QSphereMesh;
+    m_pointMesh->setRadius(0.001f);
+    pSphereTransform = new Qt3DCore::QTransform();
+    pSphereEntity->addComponent(m_pointMesh);
+    pSphereEntity->addComponent(pSphereTransform);
+    Qt3DExtras::QPhongMaterial *pSphereMat = new Qt3DExtras::QPhongMaterial;
+    pSphereMat->setAmbient(Qt::blue);
+    pSphereEntity->addComponent(pSphereMat);
+
+    // add lighting
+    QEntity *pLightRoot = new QEntity(m_pSurferRoot);
+    //Setup light positions, intensities and color
+    QList<QVector3D> lLightPositions;
+    const QColor lightColor(255,255,255);
+    const float lightIntensity = 0.2f;
+
+    lLightPositions << QVector3D(-0.5,0,0) << QVector3D(0,0,-0.5)
+                    << QVector3D(0.5,0,0) << QVector3D(0,0,0.5)
+                    << QVector3D(0,0.5,0) << QVector3D(0,-0.5,0);
+
+    //Create all the lights - make it shine
+    for(int i = 0; i < lLightPositions.size(); ++i) {
+        //Light source
+        Qt3DCore::QEntity* pLightEntity = new Qt3DCore::QEntity(pLightRoot);
+
+        Qt3DCore::QTransform* pTransform = new Qt3DCore::QTransform();
+        pTransform->setTranslation(lLightPositions.at(i));
+        pLightEntity->addComponent(pTransform);
+
+        Qt3DRender::QPointLight *pPointLight = new Qt3DRender::QPointLight(pLightEntity);
+        pPointLight->setColor(lightColor);
+        pPointLight->setIntensity(lightIntensity);
+        pLightEntity->addComponent(pPointLight);
+    }
 }
 
 
@@ -132,19 +204,9 @@ QDockWidget *Surfer::getControl()
 
 //*************************************************************************************************************
 
-// TODO check with owner ship and mdi area for garbage collection
 QWidget *Surfer::getView()
 {
-    if(!m_pView) {
-        //
-        //Pial surface
-        //
-        m_pView = new View3DSurfer();
-        m_pView->setWindowTitle(QStringLiteral("Pial surface"));
-        m_pView->setModel(m_pSurfaceModel);
-    }
-
-    return m_pView;
+    return Q_NULLPTR;
 }
 
 
@@ -170,3 +232,120 @@ void Surfer::onNewModelAvailable(QSharedPointer<AbstractModel> model)
 {
 
 }
+
+
+//*************************************************************************************************************
+
+void Surfer::onClick(QPickEvent *event)
+{
+    if(QPickTriangleEvent* tri = dynamic_cast<QPickTriangleEvent*>(event)) {
+        qDebug() << "picked triangle index: " << tri->triangleIndex();
+        qDebug() << tri->vertex1Index() << " " << tri->vertex2Index() << " " << tri->vertex3Index() << " ";
+        //qDebug() << m_surface.tris(tri->triangleIndex(), 0);
+        qDebug() << "local intersection " << tri->localIntersection();
+
+        //calculate the nearest triangle vertex to the local intersection
+//        Q_ASSERT(m_surface.tris.cols() == 3);
+//        float fMinDistance = std::numeric_limits<float>::max();
+//        int iMinVertIdx = tri->vertex1Index();
+//        for(int i = 0; i < m_surface.tris.cols(); ++i) {
+//            int iVertIdx = m_surface.tris(tri->triangleIndex(), i);
+//            float fDistance = std::sqrt(squared(m_surface.rr(iVertIdx, 0) - tri->localIntersection().x()) +
+//                                        squared(m_surface.rr(iVertIdx, 1) - tri->localIntersection().y()) +
+//                                        squared(m_surface.rr(iVertIdx, 2) - tri->localIntersection().z()));
+//            qDebug() << "Distance " << i << " : " << fDistance;
+//            if(fDistance < fMinDistance) {
+//                fMinDistance = fDistance;
+//                iMinVertIdx = iVertIdx;
+//            }
+//        }
+
+//        //Color the clicked vertex red
+//        Eigen::MatrixX3f colors;
+//        colors.resize(m_surface.rr.rows(), 3);
+//        //colors.setOnes(m_surface.rr.rows(), 3);
+//        for(int i = 0; i < colors.rows(); ++i) {
+//            colors(i, 0) = 0.2f;
+//            colors(i, 1) = 0.2f;
+//            colors(i, 2) = 0.2f;
+//        }
+//        colors(iMinVertIdx, 0) = 1.0f;
+//        colors(iMinVertIdx, 1) = 0.0f;
+//        colors(iMinVertIdx, 2) = 0.0f;
+
+//        m_mesh->setColor(colors);
+
+//        pSphereTransform->setTranslation(tri->worldIntersection());
+
+    }
+    else {
+        qDebug() << "failed";
+    }
+}
+
+
+//*************************************************************************************************************
+
+void Surfer::updateSurfaceModelMesh()
+{
+    if(m_pSurfaceModel == nullptr) {
+        return;
+    }
+
+    if(m_pSurfaceMesh == nullptr) {
+        m_pSurfaceMesh = new CustomMesh;
+    }
+
+    //TODO refactor the initialization  template with Matrix<T,Dynamic,Dynamic> ?
+    // Vertices
+    QModelIndex dataIndex = m_pSurfaceModel->index(0, 0, QModelIndex());
+    Eigen::MatrixX3f vertices(m_pSurfaceModel->columnCount(dataIndex), 3);
+
+    for(int i = 0; i < m_pSurfaceModel->columnCount(dataIndex); ++i) {
+        dataIndex = m_pSurfaceModel->index(0, i, QModelIndex());
+        Eigen::Vector3f tempVec = m_pSurfaceModel->data(dataIndex, Qt::DisplayRole).value<Eigen::Vector3f>();
+        vertices(i, 0) = tempVec[0];
+        vertices(i, 1) = tempVec[1];
+        vertices(i, 2) = tempVec[2];
+    }
+    m_pSurfaceMesh->setVertex(vertices);
+
+    // Normals
+    dataIndex = m_pSurfaceModel->index(1, 0, QModelIndex());
+    Eigen::MatrixX3f normals(m_pSurfaceModel->columnCount(dataIndex), 3);
+
+    for(int i = 0; i < m_pSurfaceModel->columnCount(dataIndex); ++i) {
+        dataIndex = m_pSurfaceModel->index(1, i, QModelIndex());
+        Eigen::Vector3f tempVec = m_pSurfaceModel->data(dataIndex, Qt::DisplayRole).value<Eigen::Vector3f>();
+        normals(i, 0) = tempVec[0];
+        normals(i, 1) = tempVec[1];
+        normals(i, 2) = tempVec[2];
+    }
+    m_pSurfaceMesh->setNormals(normals);
+
+    // Tris
+    dataIndex = m_pSurfaceModel->index(2, 0, QModelIndex());
+    Eigen::MatrixX3i tris(m_pSurfaceModel->columnCount(dataIndex), 3);
+
+    for(int i = 0; i < m_pSurfaceModel->columnCount(dataIndex); ++i) {
+        dataIndex = m_pSurfaceModel->index(2, i, QModelIndex());
+        Eigen::Vector3i tempVec = m_pSurfaceModel->data(dataIndex, Qt::DisplayRole).value<Eigen::Vector3i>();
+        tris(i, 0) = tempVec[0];
+        tris(i, 1) = tempVec[1];
+        tris(i, 2) = tempVec[2];
+    }
+    m_pSurfaceMesh->setIndex(tris);
+
+    // Colors
+    Eigen::MatrixX3f colors;
+    int rows = m_pSurfaceModel->columnCount(m_pSurfaceModel->index(0, 0, QModelIndex()));
+    colors.resize(rows, 3);
+    colors.setConstant(rows, 3, 0.6);
+    m_pSurfaceMesh->setColor(colors);
+}
+
+void Surfer::setSelectionModel(QItemSelectionModel *pSelectionModel)
+{
+    //TODO implement this
+}
+
