+++ applications/mne_analyze/extensions/surfer/surfer.pro
@@ -78,17 +78,11 @@ win32: DLLDESTDIR = $${MNE_BINARY_DIR}/mne_analyze_extensions
 unix: DESTDIR = $${MNE_BINARY_DIR}/mne_analyze_extensions
 
 SOURCES += \
-    surfer.cpp \
-    Views/baseview.cpp \
-    Views/view3danalyze.cpp \
-    Views/view3dsurfer.cpp
+    surfer.cpp
 
 HEADERS += \
     surfer_global.h \
-    surfer.h \
-    Views/baseview.h \
-    Views/view3danalyze.h \
-    Views/view3dsurfer.h
+    surfer.h
 
 INCLUDEPATH += $${EIGEN_INCLUDE_DIR}
 INCLUDEPATH += $${MNE_INCLUDE_DIR}
