+++ applications/mne_analyze/libs/anShared/Utils/types.h
@@ -85,7 +85,8 @@ namespace ANSHAREDLIB
     enum EVENT_TYPE
     {
         PING,
-        DEFAULT
+        DEFAULT,
+        EXTENSION_INIT_FINISHED
     };
 } //NAMESPACE
